Course-info-api/Course-info-cli/src/main/java/com/pluralsight/courseinfo/cli/service/PluralSightCourse.java
package com.pluralsight.courseinfo.cli.service;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;

import java.time.Duration;
import java.time.LocalTime;

@JsonIgnoreProperties(ignoreUnknown = true)
public record PluralSightCourse(String id, String title, String duration, String contentURL, boolean isRetired) {
    public long durationInMinutes(){
        return Duration.between(
                LocalTime.MIN,
                LocalTime.parse(duration())
        ).toMinutes();
    }
}
